+++ src/components/Slider/CardSkeleton.jsx
